requirements.txt
# Core LangChain (Versões v0.3 são as atuais e estáveis)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# LLM Providers
langchain-openai>=0.2.0
langchain-groq>=0.2.0

# Vector Store & Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# API & Interface
streamlit>=1.35.0
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
tiktoken>=0.7.0
pytz>=2025.2